chore: move create PR types to its own file
